Fixed the problem.
It was not in calling a builtin.
It was that an llvm basic block (deopt) didin't end with a terminator instruction.

But a new problem emerged: for some reason call to an invalid intrinsic
'llvm.sadd.with.overflow.i64.i64' is generated instead of
'llvm.sadd.with.overflow.i64'.

TODO: add verifier pass to debug build.