[safepoints]: Adopted the RewriteStatepointsForGC pass [WIP].
The code is largely based on
llvm/lib/Transforms/Scalar/RewriteStatepointsForGC.cpp.
The pass is supposed to find values which are considered pointers and
are live across the safepoint and append them to deopt paremeters of
the safepoint (thus making them appear in the generated stackmap).

This is a large piece. It compiles, but requires fixing. Plus it's
prototypal.

Conflicts:
	src/llvm/llvm-chunk.cc
	src/llvm/llvm-headers.h