[safepoints]: All patchpoint IDs are now managed in a centralized way.
Bitvectors are used to determine whether a certain patchpoint_id
corresponds to a reloc, deopt of safepoint entry.

TODO: there are calls we don't want to turn into safepoints (in deopt blocks).
This causes crashes (except --trap-on-deopt is supplied) because default
statepoint-id chosen by PlaceSafepoints is 0xabcdef00 which is larger than
int32. The idea is to change LLVM's PlaceSafepoints to recognize an attribute
indicating the call doesn't need to be transfromed to a statepoint.

Conflicts:
	src/llvm/llvm-chunk.cc