[cherry-pick] [crankshaft] Record inlined shared function infos instead of closures.
The list of inlined functions is used in exactly two places - for live
edit and to prevent code flushing for inlined functions - and those are
fine with SharedFunctionInfo and don't require a closure.

This is one additional step towards inlining based on SharedFunctionInfo
instead of JSFunction.

R=jarin@chromium.org

Review URL: https://codereview.chromium.org/1145893003

Cr-Commit-Position: refs/heads/master@{#28672}

Conflicts:
	src/arm/lithium-arm.cc
	src/debug/liveedit.cc
	src/heap/objects-visiting-inl.h
	src/lithium.cc
	src/lithium.h
	src/ppc/lithium-codegen-ppc.cc
	src/x64/lithium-x64.cc