Teaching frame walker about llvmed frames
Remember number of parameters (= number of passed arguments) of each callee
in safepoint table and use this information in IterateCompiledFrame.

This is needed, because llvm produces code which preallocates the number of
stack slots sufficient to hold locals and arguments for each callee. So we
either have to teach V8's runtime how to deal with it or teach LLVM to produce
the code in a manner V8's runtime is expecting (and I decided to go with the
former because it looks easier).

TODO:
    - I reused a 32-bit field to store the number of params.
      I think it may be too tight now, so adding another bitfield might
      be required (see the TODO in the patch).